requirements.txt
streamlit
pandas
python-dotenv
scikit-learn
